+++ CHANGELOG.md
@@ -32,6 +32,8 @@
 
 ### Fixed
 - Added missing parentheses in v-for loops
+- StatisticDashboard:
+  - Statistics and Regions are sorted alphabetically.
 
 ---
 ## 2024-06-27 v3.0.0